Add mimetypes to activity JSON data
It would be really cool to have a packagekit like popup in sugar
whenever somebody tries to open an object they don't have an activity
for.  It would also be useful with activity intents so users can
download activities if they need to open an intent.

So I have added the things in the ASLO.  I have to start this
feautre with the data I suppose.